Restructure the code to make the communication more flexible
In order to properly support the svn protocol it is necessary to have
a more flexible approach to communicating with the other end of the
link.  This is the beginning of reworking the code to allow this.  The
functions in command.py (which was commands.py) have not yet been
updated - but they do need to be.