Redesign the command system to be resumable
The command system needs to have the ability to go off and do other
things in the middle of a command - so replace the command functions
with command classes, and keep track of the current command in the
SvnRequestHandler class.  Also use metaclass and decorators to make
the syntax a bit nicer.