Basic implementation of update command
This is the absolute minimum of the update command - it only works for
an initial checkout of the entire repository.  It needs enhancing to
allows updates of the whole tree, parts of the tree and anything else
the update command may be used for.  In particular the reports from
the client are currently ignored, and the revision in the command is
supposed to be optional.