Start adding the ability to list pseudo-paths
There are parts of the Subversion url-space that don't directly map
onto something tracked by git.  For these 'pseudo-paths' we need a
mechanism for generating the data that Subversion expects to be
available.  This adds the ability to recognise that we are being asked
for such a path - but does not actually construct a reasonable reply.