Refactor the repos code to allow multiple backends
Rather than assuming that we will always be serving a git repository,
make the backends pluggable.  This means that the svn protocol
implementation can be developed on top of an svn repository - so we
don't have to worry about mapping svn concepts onto git ones while
getting the protcol implementation up and running.