Start getting log working with svn backend
Mostly this means implementing the log method for the Svn class, but
also some updates to the command module are needed as not all the
information is being passed back to the client.  This does not
complete the information being passed to the client, but it does get
log -v showing modified path and copy-from/copy-rev information.