Add checksum to get-file response
This adds the checksum to the get-file response.  It turns out that
the checksum is simply an MD5 hash of the file contents.

In order to do this we now require a reopen() method on the contents
object that is passed back.  This is made necessary by the protocol
sending the checksum before the data rather than after which means
that we need to either read it twice or cache it.  We actually do
both, if the file was able to fit into a single block of data then we
simply resend the cached data, if it was larger then we have to
re-read it.