Start moving the update infrastruture into the command
Previously we relied on the repos instance to figure out what needed
updating, and to generate the diffs.  Since all this is heavily svn
biased, and based on information provided to the command instance, it
makes more sense for most of the work to be done in the command
instance instead.  This means that we can use the ls, get_props and
get_file interface functions that are much simpler for the repos to
implement.  Also simplify the design by using recursive function
calls.