Make it possible for the caller of get_file to tell it failed
When the get_file function was called on a url/rev that didn't exist,
there was no way for the caller to tell this was the case until they
read from the contents.  Check the location first and return None for
the contents if it doesn't exist so the caller can tell immediately.