Add support for updating an existing checkout
Previously the update command would manage a fresh checkout, but would
completely fail on updating an existing checkout.  Now at least some
cases of existing checkouts are supported.  Though the communication
is nowhere near as efficient as it could be (e.g. we don't skip
updating files whose contents haven't changed, and in general no
deltas against old versions of files are computed - and we are using
svndiff0 even when we could use svndiff1).