We are going to need some supporting scripts for git interaction
We will need scripts that build the mapping information that the
interface code will use, and also scripts that dynamically update it
as the repository is updated.

This adds an initial attempt at a script that builds a mapping from a
clean repository, and also creates a place to keep the other scripts
that will be needed.