Add support for symlinks and executable bit
The executable bit and symlinks are traked using properties in
Subversion.  Add some code to set the appropriate properties from the
mode (where git stores the information).

Since the Subversion flag for symlinks is just a "parse the contents"
marker, we also need to munge the contents for symlinks, so add a
FakeData class that allows us to send generated content to the client.