Add some basic support for listing changed files
A number of things are still not covered (e.g. map change to
.gitignore to property change on containing directory, actually supply
copy paths and revs etc.), but you do at least get a list of changed
files.  Also, only commits that change the specified paths should now
be sent to the client (though we don't follow copies, which we
probably should).