Add a really simple post-receive script for updating the db
For git-svnserver to work, we need the transactions in the db to be
updated whenever the repository is updated.  The cleanest way to do
this is using the post-receive script.

This script is far too simple to deal with anything more complicated
than a simple linear update of branches.  It probably won't even
handle the creation of a new branch properly.