Actually implement the authentication stuff
Previously we supported CRAM-MD5 authentication, but we just accepted
(and then ignored) any old username provided that the password was
'test'.  Now we actually lookup the user in an authentication database
provided by the vcs backend.

In addition we also let the vcs backend know what username has been
authenticated.  Which means that we can commit using the correct user
details, rather than with those automatically picked up for the user
running the server.

Finally, we also renamed the sqlite db file from map to db, since it
contains more than just mapping informtion (and that was recorded in a
table called transactions).