Add an option to store the pid of the daemon in a file
Now you can ask git-svnserver to store the pid of the daemon process
in a file, making it easier to kill it off later.

Also tidy up the passing of options to the server - simply pass the
whole options object, rather than passing each bit of informtaion on
its own.