Start adding some tests
At the moment there is only one test - that the server actually
starts, however this is mostly about getting the beginnings of a test
framework in place.  The framework itself still needs more work, as it
doesn't provide facilities for setting up the repository to serve, or
to drive the client for us.