Add a test that git-svnserver looks like a Subversion server
We ought to be able to call 'svn info <url_of_repo>' against
git-svnserver and get a complete response back including the url we
sent.  Even against an emptry repository.

Currently this fails, as the git backend can't cope with requests that
aren't directed against a branch.  And it doesn't like not have any
data in either and returns an error instead of revision 0.