Subversion expects to be able to perform a stat on the top level
Since the is no mapping from the top level of a Subversion repository
to a real place in a git repository we have to make the results of the
stat up.  In addition, we need to be able to handle the stat of the
top-level for revision 0 where there is no "Last Changed Author" - so
we need to return an empty list for that bit of the response.