vcs: Add get_files function that returns a whole tree
Given an url and a revision get_files will return a structure
containing the properties and contents of every path to be found under
that point.  For a git repository this will hopefully be more
efficient than using ls and get_file in combination, as we can run
through a single rev-list to get the last changed data, instead of
doing it for each file individually (which git isn't very good at).