Don't try and look at the first tag when there are none left
Previously we were always trying to look at the first of the remaining
tags, even if there were none left.  Now we check that there are some
tags left to look at first - which prevents the script from crashing
when it runs out of tags.

In addition if there are any tags left after adding all the commits,
add them at the end instead of ignoring them.