Treat merges as commits
When looking for commits, we weren't treating merges as commits, which
we should have been.  In addition we were looking for 'create branch'
for branch points, when build_map is only using 'branch'.

Also remove the special case for 'delete branch', as it is the same as
the default handling.