svndiff: Make sure the client will understand zlib compressed data
In order for the client to know that data has been zlib compressed it
checks to see if the data length is less than the original data
length.  Which means that if the encoded data is as long as (or longer
than) then raw data we should just send the raw data.