vcs/git: Make log a generator function
Rather than making the user wait until we have generated all the log
entries before sending them to the client, use a generator function so
that each entry gets sent as it is ready.  This makes the user
experience much nicer.  It also means that we don't have to hold a
potentially large amount of data in memory.