Use difflib to encode diffs when we have the pre-image
If we have a pre-image for the svndiff Encoder then use difflib to
actually create some form of diff.  This isn't as complete as the
Subversion diffing (we don't use copy from target at all) but it
should decrease the amount of data that we have to send for updates
and switches.