Genereate independant file contents for use with svndiff
When we pass file contents objects to the svndiff Encoder it is going
to step through both files in chunks in parallel, which means that the
two file contents objects need to be independant - they can't have any
implicit sequential requirement, which GitObjects do as they all read
from the same cat-file instance.  We now have an IndepGitObject that
has its own cat-file instance so that it can be read in parallel with
a GitObject or anther IndepGitObject, and a way to ask for an
independant contents object - which we then use to ask for previous
contents when we a going to create a diff.