Add support for using GitFile locally
Instead of assuming that we have a commit object to give to the
GitFile tweak it so that we can use it directly by giving it a
location instead.  This allows it to be used more generally as a
wrapper for creating files in git and getting the resulting sha1.