Start adding general support for properties
Now we store all the properties that we can't map to git internals in
a JSON dumpfile for each file that has any such properties.  This
means that they are included in commits and visible to git users
(though there are issues with moving files etc. when peopl are not
aware of the property files).