Add support for the pseudo directories to check-path
We need check-path to support the pseudo directories - so that when
the client asks what kind of thing is /tags it gets the response "a
directory", etc.  Otherwise some client commands may walk themselves
up and out of the directory because they never find a valid path to
stop on.